Publication on TDAC heartbeat topic
