+++ message_bus/message_bus.py
@@ -1,5 +1,5 @@
 import subscriber
-import publisher
+from publisher import Publisher
 from heartbeat_publisher import HeartbeatPublisher
 import time
 
@@ -7,16 +7,19 @@ import time
 # This class manages input and output on the message bus
 class MessageBus():
 
-    def publish(self, topic, payload):
-        print("MessageBus publish")
-        print("  " + topic)
-        print("  " + payload)
+    publisher: Publisher
 
     def __init__(self):
         print("MessageBus.__init__")
-        self.heartbeat_publisher = HeartbeatPublisher(self)
-        self.publisher = publisher.Publisher()
+        self.publisher = Publisher(self)
+        self.heartbeat_publisher = HeartbeatPublisher(self.publisher)
         self.subscriber = subscriber.Subscriber(self)
         print("MessageBus.__init__ completed")
 
+
+    def publish(self, topic, message):
+        print("MessageBus publish")
+        print("  " + topic)
+        print("  " + message)
+        self.publisher.publish(topic, message)
