+++ message_bus/publisher.py
@@ -3,17 +3,9 @@ import paho.mqtt.client as mqtt
 # https://www.eclipse.org/paho/index.php?page=clients/python/docs/index.php
 
 
-class Publisher:
-    def __init__(self):
-        print("Publisher.__init__")
-        self.publisher_client = PublisherClient(self)
-
-    def publish(self, topic, message):
-        self.publisher_client.publish(topic, message)
-
 
 # non-asynchronous MQTT publisher
-class PublisherClient:
+class Publisher:
 
     # TODO pass these in
     host = "localhost"
@@ -21,17 +13,15 @@ class PublisherClient:
     keepalive = 6000
     bind_address = ""
 
-    client: mqtt.Client
-
-    def __init__(self, publisher):
-        print("PublisherClient.__init__")
-        self.publisher = publisher
+    def __init__(self, message_bus):
+        print("Publisher.__init__")
+        self.message_bus = message_bus
         self.client = mqtt.Client()
         print("Publisher connecting to Message Bus... ")
         self.client.connect(self.host, self.port, self.keepalive, self.bind_address)
         self.client.loop_start()
 
     def publish(self, topic, message):
-        print("PublisherClient.publish(" + topic + ", " + message + ")")
+        print("Publisher.publish(" + topic + ", " + message + ")")
         self.client.publish(topic, message)
